pyproject.toml
[tool.poetry]
name = "pypractice"
version = "0.1.0"
description = ""
authors = ["alsktv <alsktv@naver.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.4"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
pyjwt = "^2.8.0"
django-environ = "^0.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
